requirements.txt
attrs==23.1.0
aws-wsgi==0.2.7
bleach==6.1.0
blinker==1.6.3
# Editable Git install with no remote (bloggy==0.0.1)
-e ./
boto3==1.28.62
botocore==1.31.62
cattrs==23.1.2
click==8.1.7
colorama==0.4.6
coverage==7.3.2
docopt==0.6.2
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.8.0
Flask==3.0.0
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Markdown==3.5
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
pytest-watch==4.2.0
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==1.26.17
watchdog==3.0.0
webencodings==0.5.1
Werkzeug==3.0.0
zipp==3.17.0
